refactor: プロフィール機能をfeatureディレクトリに集約
