+++ app/web/app/.server/actions/index.ts
@@ -2,4 +2,3 @@ export * from "./eventAction";
 export * from "./friendAction";
 export * from "./memberAction";
 export * from "./paymentAction";
-export * from "./profileAction";
