+++ app/web/app/features/profile/actions/index.ts
@@ -1,0 +1,1 @@
+export * from "./update-profile-action";
