+++ app/web/app/features/profile/loaders/index.ts
@@ -1,0 +1,1 @@
+export * from "./profile-loader";
