+++ app/web/app/features/profile/schemas/index.ts
@@ -3,7 +3,9 @@ import type { UserEndpoints_UserPutRequest } from "~/api/@types";
 
 import type { ToZod } from "~/lib/toZod";
 
-export const profileFormSchema = z.object<ToZod<UserEndpoints_UserPutRequest>>({
+export const updateProfileSchema = z.object<
+	ToZod<UserEndpoints_UserPutRequest>
+>({
 	name: z.string().min(1, {
 		message: "ユーザー名を入力してください",
 	}),
